+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+**v0.3.0:** **⚠️ BREAKING CHANGE**
+- **Breaking change**: IDs change. Algorithm has been fine-tuned for better performance [[Issue #11](https://github.com/sqids/sqids-spec/issues/11)]
+- `alphabet` cannot contain multibyte characters
+- `min_length` was changed from `usize` to `u8`
+- Max blocklist re-encoding attempts has been capped at the length of the alphabet - 1
+- Minimum alphabet length has changed from 5 to 3
+- `min_value()` and `max_value()` functions have been removed
+
 **v0.2.1:**
 - Bug fix: spec update (PR #7): blocklist filtering in uppercase-only alphabet [[PR #7](https://github.com/sqids/sqids-spec/pull/7)]
 - Updating Github Actions to use stable toolchain instead of nightly